use a different algorithm for Day 2 Part 2
In order to find the pair of numbers that evenly divide, instead of testing for
all multiples of one number whether the other number is in the set, simply try
for all pairs of numbers in the set whether they evenly divide.

In the algorithmic complexity, this should replace a factor "ratio between
largest and smallest number in the set" by a factor "size of the set".