Try to improve day 8 part 2
The idea is to build a lookup table for how to reach the xxZ positions
given the current position and index in the instructions list.

Then simulate the movement of the ghosts by advancing the ghost that has
taken the least amount of steps so far, until all of them are on a xxZ
position and have the same number of steps.

But even with the LUT this will take way too long (after knowing the
correct answer and printing the progress, probably several hours).